[ISSUE #9623]♻️Extract Tokio server inline tests into a sibling module
